This bugfix fixes: http://intranet/bugzilla/show_bug.cgi?id=18
Prior to this fix a target directory for the installation was created immediately when 
the installer was started - even though the user could change it later on. This 
bugfixes solves this issue by delaying the creation of the directory until the user has 
picked a target directory and started the installation.