Initial support for extension bundles.
This includes:
The ability to install extension bundles and these will be added to the system class path on restart.
The ability to load classes using the system class loader or the framework's class loader.

Unfortunately this excludes:
The system bundle's exports are NOT updated
Can not search for resources in extensions